refactor: model 패키지 하위 구조를 cart 및 product로 분리 및 재구성
