+++ app/src/main/java/woowacourse/shopping/model/product/Money.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.model.product
 
 @JvmInline
 value class Money(
