+++ app/src/main/java/woowacourse/shopping/model/product/Product.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.model.product
 
 data class Product(
     val id: Long,
