+++ app/src/main/java/woowacourse/shopping/model/product/Products.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.model.product
 
 class Products(
     products: List<Product>,
