+++ app/src/main/java/woowacourse/shopping/model/product/RecentProduct.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.model.product
 
 data class RecentProduct(
     val productId: Long,
