+++ app/src/main/java/woowacourse/shopping/repository/CartRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.repository
 
-import woowacourse.shopping.model.CartItem
+import woowacourse.shopping.model.cart.CartItem
 import woowacourse.shopping.repository.query.CartPageResult
 
 interface CartRepository {
