+++ app/src/main/java/woowacourse/shopping/repository/ProductRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.repository
 
-import woowacourse.shopping.model.Product
-import woowacourse.shopping.model.Products
+import woowacourse.shopping.model.product.Product
+import woowacourse.shopping.model.product.Products
 
 interface ProductRepository {
     val size: Int
