+++ app/src/main/java/woowacourse/shopping/repository/RecentProductRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.repository
 
-import woowacourse.shopping.model.RecentProduct
+import woowacourse.shopping.model.product.RecentProduct
 
 interface RecentProductRepository {
     suspend fun recordView(productId: Long)
