+++ app/src/main/java/woowacourse/shopping/ui/common/component/recentlyviewed/RecentlyViewedItem.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import woowacourse.shopping.R
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 import woowacourse.shopping.ui.ShoppingTypography
 import woowacourse.shopping.ui.fixture.MockProducts
