+++ app/src/main/java/woowacourse/shopping/ui/productdetail/ProductDetailUiState.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.ui.productdetail
 
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 
 data class ProductDetailUiState(
     val product: Product? = null,
