+++ app/src/main/java/woowacourse/shopping/ui/shopping/ShoppingProductUiState.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.ui.shopping
 
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 
 data class ShoppingProductUiState(
     val product: Product,
