+++ app/src/main/java/woowacourse/shopping/ui/shopping/ShoppingUiState.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.ui.shopping
 
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 
 data class ShoppingUiState(
     val productListState: ProductListUiState = ProductListUiState.Loading,
