+++ app/src/main/java/woowacourse/shopping/ui/shopping/component/ShoppingBody.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 import woowacourse.shopping.ui.common.component.recentlyviewed.RecentlyViewedSection
 import woowacourse.shopping.ui.fixture.MockProducts
 import woowacourse.shopping.ui.shopping.ProductListUiState
