+++ app/src/test/java/woowacourse/shopping/model/MoneyTest.kt
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import woowacourse.shopping.model.product.Money
 
 class MoneyTest {
     @Test
