+++ app/src/test/java/woowacourse/shopping/repository/CartRepositoryFixture.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.repository
 
-import woowacourse.shopping.model.Money
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Money
+import woowacourse.shopping.model.product.Product
 
 object CartRepositoryFixture {
     val shrimpCracker =
