+++ app/src/test/java/woowacourse/shopping/repository/FakeCartRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.repository
 
-import woowacourse.shopping.model.Cart
-import woowacourse.shopping.model.CartItem
+import woowacourse.shopping.model.cart.Cart
+import woowacourse.shopping.model.cart.CartItem
 import woowacourse.shopping.repository.query.CartPageResult
 
 class FakeCartRepository : CartRepository {
